[ESC-29] 전체 QA 1차 - 2 (#63)
* [ESC-29] 기부참여를 마치고 홈에 돌아왔을때 다시 기부하러가기 버튼을 클릭하면 토스트가 노출되어야 하는데 그대로 종이비행기 선택페이지로 이동
- 참여여부 확인을 앱에 진입한 최초에만 하기 때문
- 참여여부 확인을 main 페이지에 들어올때 마다 하도록 변경

* [ESC-29] 기부내역/상세 페이지 네비게이션 오른쪽 상단에 홈 버튼 추가: 기존에는 해당 페이지에서 홈으로 갈 수 있는 방법이 없었음
